Successfully got the torrented files to move, and there is there is evidence of clean up
This was not done in an ideal fashion, as I would have liked to place
all the files in the root of the author's directory (I decided to get
rid of the "raw" directory across the board).

Because Rust's fs::rename() method requires that the destination be a
directory which exists and is empty. My previous plan would have failed,
as the "raw" directory would never be empty (under current conditions).
So the workaround was to ensure that torrented books be moved to new
directories that are named after them.

As a consequence of this choice, the chunking and other text processing
tasks will have to be written in a way that acknowledges this new
directory structure when grabbing text.